Switch permission to use TreeConnect.MaximalAccess
Previously share permissions would be binary READ or none and there was
no testing for write permissions. This was done by calling
`SMBConnection.listPath` for each share and seeing if there was an
error. Now, share permissions come from Sending a TreeConnect packet for
each share and reporting the returned MaximalAccess bitmask.

There are some disadvantages to this:
- Might not work with SMBv1
- Gives false positives when a share has permissions but the root folder
  has no permissions.
- This code is not well tested

But also major advantages:
- Reports granular permissions including write access.
- Sends fewer packets and should be faster.

This is a testing commit.